# docs/openapi.json
{"openapi":"3.0.2","info":{"title":"Acleda Map 🔥","version":"1.0.0"},"paths":{"/api/v1/merchant":{"post":{"tags":["merchant"],"summary":"Create Merchant","operationId":"create_merchant_api_v1_merchant_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMerchant"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShowMerchant"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/merchants":{"get":{"tags":["merchant"],"summary":"Get All Merchant","operationId":"get_all_merchant_api_v1_merchants_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Merchant Api V1 Merchants Get","type":"array","items":{"$ref":"#/components/schemas/ShowMerchant"}}}}}}}},"/api/v1/merchants/pagination":{"get":{"tags":["merchant"],"summary":"Get Merchants With Pagination","operationId":"get_merchants_with_pagination_api_v1_merchants_pagination_get","parameters":[{"required":true,"schema":{"title":"Page","type":"integer"},"name":"page","in":"query"},{"required":true,"schema":{"title":"Per Page","type":"integer"},"name":"per_page","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Merchants With Pagination Api V1 Merchants Pagination Get","type":"array","items":{"$ref":"#/components/schemas/ShowMerchant"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/merchant/{id}":{"get":{"tags":["merchant"],"summary":"Get Merchant","operationId":"get_merchant_api_v1_merchant__id__get","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShowMerchant"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/merchant/{name}":{"get":{"tags":["merchant"],"summary":"Get Merchant By Name","operationId":"get_merchant_by_name_api_v1_merchant__name__get","parameters":[{"required":true,"schema":{"title":"Name","type":"string"},"name":"name","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShowMerchant"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/merchants/{lat}/{long}":{"get":{"tags":["merchant"],"summary":"Get Merchant By Lat Long","operationId":"get_merchant_by_lat_long_api_v1_merchants__lat___long__get","parameters":[{"required":true,"schema":{"title":"Lat","type":"number"},"name":"lat","in":"path"},{"required":true,"schema":{"title":"Long","type":"number"},"name":"long","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShowMerchant"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/merchants/top10/{lat}/{long}":{"get":{"tags":["merchant"],"summary":"Get Top10 Nearest Merchant","operationId":"get_top10_nearest_merchant_api_v1_merchants_top10__lat___long__get","parameters":[{"required":true,"schema":{"title":"Lat","type":"number"},"name":"lat","in":"path"},{"required":true,"schema":{"title":"Long","type":"number"},"name":"long","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Top10 Nearest Merchant Api V1 Merchants Top10  Lat   Long  Get","type":"array","items":{"$ref":"#/components/schemas/ShowMerchant"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Root"],"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CreateMerchant":{"title":"CreateMerchant","required":["id","name","latitude","longitude"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ShowMerchant":{"title":"ShowMerchant","required":["id"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}